v3.2.18-rc4
adding user secrets, updating callback imports, tweaking socks

+++ mythic-docker/src/database/structs/Callbackport.go
@@ -10,6 +10,6 @@ type Callbackport struct {
 	RemotePort    int    `db:"remote_port"`
 	RemoteIP      string `db:"remote_ip"`
 	Deleted       bool   `db:"deleted"`
-	BytesReceived int    `db:"bytes_received"`
-	BytesSent     int    `db:"bytes_sent"`
+	BytesReceived int64  `db:"bytes_received"`
+	BytesSent     int64  `db:"bytes_sent"`
 }
